<CommandTree> managing state.
